requirements.txt
flask==3.0.0
opensearch-py==2.3.2
boto3==1.28.62
python-dotenv==1.0.0
flask-cors==4.0.0 